Reorganise referral referrer form objects and views
